Fixes to the space reclaiming protocols/services
GUI: Add spinbox to enter desired max backup size

+++ Peer/Program.cs
@@ -63,7 +63,7 @@ namespace Peer
                 return;
             }
 
-            int maxBackupSize = Config.Global.DiskSpace; // Max site used to backup (locally)
+            long maxBackupSize = Config.Global.DiskSpace; // Max site used to backup (locally)
             int chunkSize = Config.Global.ChunkSize; // Size of each chunk stored locally and sent over the network
 
             // Backup chunk protocol configurations
